migrated GitRebaserTest to Junit 5, with workaround for tmpDirectory

+++ ci-droid-tasks-consumer-infrastructure/src/test/java/com/societegenerale/cidroid/tasks/consumer/infrastructure/GitRebaserTest.java
@@ -12,17 +12,16 @@ import org.eclipse.jgit.api.*;
 import org.eclipse.jgit.api.errors.GitAPIException;
 import org.eclipse.jgit.lib.*;
 import org.eclipse.jgit.revwalk.RevCommit;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.TemporaryFolder;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.time.Instant;
 import java.util.HashSet;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
@@ -37,8 +36,7 @@ import static org.mockito.internal.verification.VerificationModeFactory.times;
 
 public class GitRebaserTest {
 
-    @Rule
-    public TemporaryFolder tmpFolder = new TemporaryFolder();
+    static String tmpFolderName="."+File.separator+"target"+File.separator+"GitRebaserTest-tmpDir-"+ Instant.now().toEpochMilli() ;
 
     GitWrapper mockGitWrapper=mock(GitWrapper.class);
 
@@ -58,7 +56,7 @@ public class GitRebaserTest {
 
     CheckoutCommand mockCheckoutCommand = mock(CheckoutCommand.class);
 
-    @Before
+    @BeforeEach
     public void setUp() throws IOException, GitAPIException {
 
         File tmpDirectory = createWorkingDirIfRequired();
@@ -125,7 +123,7 @@ public class GitRebaserTest {
 
     }
 
-    @After
+    @AfterEach
     public void assertGitIsClosed() throws GitAPIException {
 
         verify(mockGit, times(1)).close();
@@ -335,11 +333,17 @@ public class GitRebaserTest {
     }
 
     /**
-     * Impossible to mock RevCommit with Mockito (some methods are final), so non choice but build a tmp Git repo to build RevCommit instances
+     * Impossible to mock RevCommit with Mockito (some methods are final), so no choice but build a tmp Git repo to build RevCommit instances
      */
     private RevCommit buildRevCommit(String fileName, String content) throws GitAPIException, IOException {
 
-        Git git = Git.init().setDirectory(tmpFolder.getRoot()).call();
+        File tmpFolder=new File(tmpFolderName);
+
+        if(!tmpFolder.exists() && !tmpFolder.mkdirs()){
+            throw new RuntimeException("couldn't create tmp dir");
+        }
+
+        Git git = Git.init().setDirectory(tmpFolder).call();
 
         File file = new File(git.getRepository().getWorkTree(), fileName);
         try (FileOutputStream os = new FileOutputStream(file)) {
